requirements.txt
fastapi
uvicorn
playwright
openai
python-dotenv
fastapi
uvicorn
playwright
openai
python-dotenv
celery
fastapi
uvicorn
playwright
openai
python-dotenv
fastapi
uvicorn
playwright
openai
python-dotenv
celery
fastapi
uvicorn
playwright
openai
python-dotenv
fastapi
uvicorn
playwright
openai
python-dotenv
celery
fastapi
uvicorn
playwright
openai
python-dotenv
fastapi
uvicorn
playwright
openai
python-dotenv
celery
redis
asyncpg
requests
python-multipart
twilio
google-generativeai
passlib[bcrypt]
python-jose[cryptography]